Add base Filesystem testcase
